Trading working uncap indicators for working party creation
We start storing the party ID here so that we know (and can come to know) when we are working with an existing party.

Updating the weapon grid based on `weapons` the state vs `props.weapons` makes saving items to a party work properly.

Unfortunately, it sends us back to square one with debouncing the uncap indicators.